Create signed tags (#12)
* Use Git CLI for tag creation

* Git config stuff

* Copy a gnupg directory into the archive

* Make signed tags work

...And some other stuff.

- Store required data (app key, GPG stuff) on S3 instead of in the bundle
- Tweak some internal APIs (return error instead of string)
- Use pkg/errors for wrapping instead of ugly string concatenation
- More things that I don't feel like reading through the massive diff for

* Skip tag creation if one already exists

* Remove unneeded gitignore entries

* Check for some env variables at startup

* We don't actually need S3

Using tar lets us preserve file permission for the GNUPG stuff.

* Move some code around, delete some stuff

Now that there's no network involved, it's cheap and risk-free to do
the setup steps at initialization instead of waiting until we need
them.

* Delete any leftover GPG agent socket

* Change some spelling

* Rename some more things

* Remove an unused variable, add more detailed comment

+++ build.sh
@@ -1,11 +1,26 @@
 #!/usr/bin/env bash
 
 FNS=("github")
+PEM="tag-bot.pem"
+GPG="gnupg"
+RESOURCES="bin/resources.tar"
+
+if [ ! -f  "$PEM" ]; then
+    echo "File $PEM does not exist"
+    exit 1
+fi
+if [ ! -d  "$GPG" ]; then
+    echo "Directory $GPG does not exist"
+    exit 1
+fi
 
 rm -rf bin
 mkdir bin
-cp *.pem bin
-chmod 644 bin/*.pem
+chmod 644 "$PEM"
+rm -f "$GPG/S.gpg-agent"
+find "$GPG" -type d -exec chmod 700 {} \;
+find "$GPG" -type f -exec chmod 600 {} \;
+tar -cf "$RESOURCES" "$GPG" "$PEM"
 
 for fn in "${FNS[@]}"; do
   (
